+++ magnolia/src/editor/panels/materials_panel.hpp
@@ -8,6 +8,6 @@ namespace mag
     class MaterialsPanel
     {
         public:
-            void render(const ImGuiWindowFlags window_flags, ECS& ecs, const u64 selected_entity_id);
+            void render(const ImGuiWindowFlags window_flags, ECS& ecs, const u32 selected_entity_id);
     };
 };  // namespace mag
