+++ magnolia/src/editor/panels/properties_panel.hpp
@@ -8,6 +8,6 @@ namespace mag
     class PropertiesPanel
     {
         public:
-            void render(const ImGuiWindowFlags window_flags, ECS& ecs, const u64 selected_entity_id);
+            void render(const ImGuiWindowFlags window_flags, ECS& ecs, const u32 selected_entity_id);
     };
 };  // namespace mag
